requirements.txt
streamlit
numpy
Pillow
insightface
gdown
onnxruntime
opencv-python-headless
